+++ ot/da.py
@@ -598,23 +598,24 @@ class BaseTransport(BaseEstimator):
                 if self.limit_max != np.inf:
                     self.limit_max = self.limit_max * nx.max(self.cost_)
 
-                # missing_labels is a (ns, nt) matrix of {0, 1} such that
-                # the cells (i, j) has 0 iff either ys[i] or yt[j] is masked
-                missing_ys = (ys == -1) + nx.zeros(ys.shape, type_as=ys)
-                missing_yt = (yt == -1) + nx.zeros(yt.shape, type_as=yt)
-                missing_labels = missing_ys[:, None] @ missing_yt[None, :]
-                # labels_match is a (ns, nt) matrix of {True, False} such that
-                # the cells (i, j) has False if ys[i] != yt[i]
-                label_match = (ys[:, None] - yt[None, :]) != 0
+                # present_labels is a (ns, nt) matrix of {0, 1} such that
+                # the cell (i, j) is 1 iff both ys[i] and yt[j] are labeled
+                # (i.e. neither is masked with -1)
+                present_ys = (ys != -1) + nx.zeros(ys.shape, type_as=ys)
+                present_yt = (yt != -1) + nx.zeros(yt.shape, type_as=yt)
+                present_labels = present_ys[:, None] @ present_yt[None, :]
+                # label_mismatch is a (ns, nt) matrix of {True, False} such that
+                # the cell (i, j) is True if ys[i] != yt[j]
+                label_mismatch = (ys[:, None] - yt[None, :]) != 0
                 # cost correction is a (ns, nt) matrix of {-Inf, float, Inf} such
                 # that he cells (i, j) has -Inf where there's no correction necessary
-                # by 'correction' we mean setting cost to a large value when
-                # labels do not match
+                # by 'correction' we mean setting cost to a large value when two
+                # labeled samples have different labels
                 # we suppress potential RuntimeWarning caused by Inf multiplication
                 # (as we explicitly cover potential NANs later)
                 with warnings.catch_warnings():
                     warnings.simplefilter("ignore", category=RuntimeWarning)
-                    cost_correction = label_match * missing_labels * self.limit_max
+                    cost_correction = label_mismatch * present_labels * self.limit_max
                 # this operation is necessary because 0 * Inf = NAN
                 # thus is irrelevant when limit_max is finite
                 cost_correction = nx.nan_to_num(cost_correction, -np.inf)